requirements.txt
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
moviepy==1.0.3
Pillow>=10.0.0
pygame>=2.0.0

